RAG Project with Langchain and OepnAI/requirements.txt
langchain==0.3.27
langchain-openai==0.3.28
langchain-core==0.3.72
langchain-chroma==0.2.5
langchain-community
langchain-text-splitters==0.3.9
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
requests==2.32.4
python-dotenv==1.1.1
python-multipart==0.0.20
beautifulsoup4==4.13.4
docx2txt
pypdf